feat: move HTMX fallback script to external file
- Moved inline HTMX fallback script from `app/layout.tsx` to `public/htmx-fallback.js`
- Loaded the script using `next/script` with `strategy="beforeInteractive"` to improve security and avoid CSP issues with inline scripts
- Maintained the same logic for setting up safety fallback handlers for HTMX events
