+++ public/htmx-fallback.js
@@ -1,0 +1,15 @@
+(function() {
+  const htmxEvents = [
+    'sseError', 'sseOpen', 'swapError', 'targetError', 'timeout',
+    'validation:validate', 'validation:failed', 'validation:halted',
+    'xhr:abort', 'xhr:loadend', 'xhr:loadstart'
+  ];
+  htmxEvents.forEach(event => {
+    const funcName = 'func ' + event;
+    if (typeof window[funcName] === 'undefined') {
+      window[funcName] = function() {
+        console.warn('HTMX event handler "' + funcName + '" was called but not defined. Providing safety fallback.');
+      };
+    }
+  });
+})();
